Did it another way and this is much much better. No need to merge yet as I will complete the entire remaining endpoints before need to merge.
Request Daniel to verify if this method is good enough and then will proceed with the rest of the endpoints.

Since Json attributes and ShouldSerialize<PropName>() methods are used the need for the PutJsonString is gone so this code is removed.

Though personally I prefer PutJsonString for everything but I am old school. However for return values definitely support classes to be returned for end user to use of course of course.